Provides cross-compilation Docker image
This commit provides a cross compilation Docker image which compiles the
project using the same scripts as it is used on the remote machine.

With the cross compilation image it is possible to test it locally.

We introduced a new compiler on the build step, so we are capable of
using `cgo` during compilation, as it was [reported] to be required for
newer versions of Nomad.

This has been tested running Docker from Mac, and sending the output to
a Pi Zero W using `scp`. The binary starts up, but there were no further
tests on how the scheduler behaves at this moment.

This also has been tested compiling Nomad `0.8.7` and `master`
(currently `0.9.0-dev`), using the `CGO_ENABLED` and the correct
gcc `armv6l` compiler.

[reported]: https://github.com/hashicorp/nomad/pull/5337